Providing a "printable" version of `PathingMap` to avoid `Show` conflict
As I write this, I'm on a plane, so I can't really look up a better way
to do this, but surely I should fix this situation later.  I think it
would be dumb if I couldn't more-seamlessly override the `Show` for 2D
arrays whenever I have a declared `PathingMap`.