Fixing `Interpreter` to flip the input's axes correctly
This is something that I've always found tricky, and I assumed that I
messed it up when I wrote the first implementation here in Haskell.

In the Scala version, the transformation is done by reassigning the
contents of arrays, but Haskell doesn't really allow that (and I'm
fine with that).

Solving it--like most things in Haskell--mostly just involved sitting
down and thinking about what the problem was that I wanted to solve.